requirements.txt
aiogram>=3.0.0
langchain-gigachat>=0.2.0
langchain-core>=0.1.0
gigachat>=0.1.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
redis>=5.0.0
python-dotenv>=1.0.0
